refactor!: rename global config to common config
Change all references to `global` config to `common` config. This is to prevent confusion with the to-be introduced `global` config.

Note that this is a breaking change in the Wasm plugin API, which has therefore been bumped to schema version 5. The process plugin API is luckily unchanged.

BREAKING CHANGE: rename any identifier refering to the `global` config (`global_config`, `GlobalConfig`, `config.global`, …) to equivalent `common` config
BREAKING CHANGE: bump Wasm plugin API schema version to 5